Removing some unfinished or unused classes.
What might be salvageable:
IdentifierRegexFilter - a filter similar to IdentifierFilter, but it
passes on a match with a given regex as opposed to equality to a String.
PathFilter - I think it's supposed to filter every level of a tree with
different given subfilters.
treeview2 - A different version of treeview that supports models in some
way.